refactor(schedules): review feedback — naming, placement, dead config
DatabaseSchedule moves out of Platform/Tasks into Appwrite\Schedule: it is
not a task, and "ScheduleSource" both repeated the context it sat in and
named itself after the library's interface rather than what it is — the
schedules collection, read as a schedule source.

The combined task initialises the three imperatively. The array plus
foreach bought nothing at three items, and the comment above it said what
the code says.

ENQUEUE_LOOKBACK goes: 300 seconds is the library default, so declaring it
was three constants and three arguments that changed nothing.
