+++ src/Appwrite/Platform/Tasks/ScheduleFunctions.php
@@ -4,6 +4,7 @@ namespace Appwrite\Platform\Tasks;
 
 use Appwrite\Event\Message\Func as FunctionMessage;
 use Appwrite\Event\Publisher\Func as FunctionPublisher;
+use Appwrite\Schedule\DatabaseSchedule;
 use Utopia\Database\Database;
 use Utopia\Database\Document;
 use Utopia\Platform\Action;
@@ -23,7 +24,6 @@ class ScheduleFunctions extends Action
     public const UPDATE_TIMER = 10; // seconds between reconciliations
     public const ENQUEUE_TIMER = 60; // seconds between ticks
     public const ENQUEUE_LOOKAHEAD = 60; // seconds of lead time, so a dispatch can sleep to the second
-    public const ENQUEUE_LOOKBACK = 300; // seconds of missed runs a restart recovers
 
     public function __construct()
     {
@@ -76,7 +76,7 @@ class ScheduleFunctions extends Action
     {
         Span::init('schedule.functions.boot');
 
-        $source = new ScheduleSource(
+        $source = new DatabaseSchedule(
             dbForPlatform: $dbForPlatform,
             getProjectDB: $getProjectDB,
             isResourceBlocked: $getIsResourceBlocked,
@@ -100,7 +100,6 @@ class ScheduleFunctions extends Action
             tickSeconds: self::ENQUEUE_TIMER,
             syncSeconds: self::UPDATE_TIMER,
             leadSeconds: self::ENQUEUE_LOOKAHEAD,
-            recoverSeconds: self::ENQUEUE_LOOKBACK,
             telemetry: $telemetry,
             onError: function (\Throwable $error): void {
                 Span::init('schedule.functions.reconcile');
@@ -118,7 +117,7 @@ class ScheduleFunctions extends Action
 
     protected function updateProjectAccess(Document $project, Database $dbForPlatform): void
     {
-        ScheduleSource::touchProject($project, $dbForPlatform);
+        DatabaseSchedule::touchProject($project, $dbForPlatform);
     }
 
     /**
